Wifi: Read current BSSID directly from NetworkInfo
In QNS, currentBssid variable holds the connected BSSID. This variable is
currently updated only when it is NULL (is set to NULL only when
disconnected). Thus, the current check of updating the currentBSSID only
when NULL would miss the "current BSSID" on every firmware based roaming
and since the current BSSID is not updated, it shall result in to improper
QNS behavior for giving wrong score to the "current BSSID". Thus, ensure
that the currentBSSID is always read rather than only reading when NULL.

CRs-Fixed: 1091343
Change-Id: I9c39757a0b8b6784b2d9b8882555b4d069c05e4f